+++ CHANGELOG.md
@@ -1311,3 +1311,4 @@
 - [2026-05-20 22:06] feat(core): implement responsive core view component
 - [2026-05-22 10:32] chore(test): update dependencies and cleanup unused imports in test
 - [2026-05-22 10:41] chore(analytics): update dependencies and cleanup unused imports in analytics
+- [2026-05-22 10:44] refactor(security): migrate security controller to modern notifier syntax